src/test/kotlin/com/github/lemfi/adventofcode/year2021/Day10Test.kt
package com.github.lemfi.adventofcode.year2021

import org.junit.jupiter.api.Assertions
import org.junit.jupiter.api.Test

class Day10Test {

    @Test
    fun star1() {
        Assertions.assertEquals(
            26397, Day10.star1(
                """[({(<(())[]>[[{[]{<()<>>
[(()[<>])]({[<{<<[]>>(
{([(<{}[<>[]}>{[]{[(<()>
(((({<>}<{<{<>}{[]{[]{}
[[<[([]))<([[{}[[()]]]
[{[{({}]{}}([{[{{{}}([]
{<[[]]>}<{[{[{[]{()[[[]
[<(<(<(<{}))><([]([]()
<{([([[(<>()){}]>(<<{{
<{([{{}}[<[[[<>{}]]]>[]]"""
            )
        )
    }

    @Test
    fun star2() {
        Assertions.assertEquals(
            288957, Day10.star2(
                """[({(<(())[]>[[{[]{<()<>>
[(()[<>])]({[<{<<[]>>(
{([(<{}[<>[]}>{[]{[(<()>
(((({<>}<{<{<>}{[]{[]{}
[[<[([]))<([[{}[[()]]]
[{[{({}]{}}([{[{{{}}([]
{<[[]]>}<{[{[{[]{()[[[]
[<(<(<(<{}))><([]([]()
<{([([[(<>()){}]>(<<{{
<{([{{}}[<[[[<>{}]]]>[]]"""
            )
        )
    }

    @Test
    fun puzzle1() {
        Assertions.assertEquals(271245, Day10.star1(null))
    }

    @Test
    fun puzzle2() {
        Assertions.assertEquals(1685293086, Day10.star2(null))
    }
}